# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: ShellWatch HTTP API
  version: "1.1.0"
  description: >-
    Passkey-backed SSH brokering for humans and AI agents. All authenticated
    surfaces present a Hydra-issued opaque OAuth bearer token, validated by
    introspection; the token `sub` is the account id. Authorization is by scope
    (`ui` for /api + /ws, `mcp` for /mcp, `agent` for /agent-proxy).
  license:
    name: FSL-1.1-Apache-2.0
    url: https://fsl.software/
servers:
  - url: "{externalUrl}"
    description: Deployment base URL (config.server.externalUrl)
    variables:
      externalUrl:
        default: http://localhost:3000
...
security:
  - bearerUi: []
components:
  securitySchemes:
    bearerUi:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        Hydra opaque access token with the `ui` scope. Required for all `/api/*`
        routes except those listed as `security: []`. For `/ws` the same token
        is passed via `Sec-WebSocket-Protocol: shellwatch.bearer, <token>`
        because browsers cannot set an Authorization header on a WS handshake.
    bearerMcp:
      type: http
      scheme: bearer
      description: Opaque token with `mcp` scope (guards /mcp; IP-allowlisted).
    bearerAgent:
      type: http
      scheme: bearer
      description: Opaque token with `agent` scope (guards /agent-proxy, when enabled).
...
  schemas:
...
    Error:
      type: object
      description: Standard error envelope. Almost every 4xx/5xx returns this.
      required: [error]
      properties:
        error: { type: string }
      additionalProperties: true
...
    SessionLifecyclePage:
      type: object
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/SessionLifecycleRow" }
        nextCursor:
          type: string
          nullable: true
          description: Opaque cursor; null on the final page.
      required: [rows, nextCursor]
    SessionLifecycleRow:
      type: object
      properties:
        sessionId: { type: string }
        accountId: { type: string }
        endpointId: { type: string }
        # Enforced by DB CHECK constraints (src/db/schema.ts):
        source: { type: string, enum: [ui, mcp, ssh] }
        status: { type: string, enum: [open, closed, error] }
        createdAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time, nullable: true }
        durationMs: { type: integer, nullable: true }
        sourceIp: { type: string, nullable: true }
        mcpReason: { type: string, nullable: true }
        mcpClientName: { type: string, nullable: true }
        mcpClientVersion: { type: string, nullable: true }
        clientHostname: { type: string, nullable: true }
        clientOs: { type: string, nullable: true }
        clientVersion: { type: string, nullable: true }
        closeReason: { type: string, nullable: true }
    SigningRequestPage:
      type: object
      properties:
        rows:
          type: array
          items: { $ref: "#/components/schemas/SigningRequestRow" }
        nextCursor: { type: string, nullable: true }
      required: [rows, nextCursor]
    SigningRequestRow:
      type: object
      properties:
        id: { type: string }
        accountId: { type: string }
        type: { type: string }
        source: { type: string, enum: [endpoint-auth, agent-forwarding, agent-proxy] }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
        outcome:
          type: string
          nullable: true
          enum: [approved, denied, expired, cancelled, null]
        latencyMs: { type: integer, nullable: true }
        sourceIp: { type: string, nullable: true }
        endpointLabel: { type: string, nullable: true }
        endpointAddress: { type: string, nullable: true }
        sessionId: { type: string, nullable: true }
        mcpReason: { type: string, nullable: true }
        mcpClientName: { type: string, nullable: true }
        mcpClientVersion: { type: string, nullable: true }
        clientHostname: { type: string, nullable: true }
        clientOs: { type: string, nullable: true }
        clientVersion: { type: string, nullable: true }
        credentialId: { type: string, nullable: true }
        passkeyLabel: { type: string, nullable: true }
        userVerification: { type: string, nullable: true }
        keyLabel: { type: string, nullable: true }
        keyFingerprint: { type: string, nullable: true }
        cancelReason: { type: string, nullable: true }
...
paths:
...
  /api/hydra/register:
    post:
      operationId: registerOauthClient
      tags: [discovery]
      security: []
      summary: Mediated Dynamic Client Registration (RFC 7591)
      description: >-
        Policy enforced locally (redirect-uri patterns, allowed scopes); the
        client is minted in Hydra as a public PKCE client. Rate-limited
        (10 / 15 min). `offline_access` is always added to the granted scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                redirect_uris: { type: array, items: { type: string } }
                scope: { type: string, description: "space-delimited; subset of allowed" }
                client_name: { type: string, default: "MCP Client" }
              required: [redirect_uris]
      responses:
        "201":
          description: Registered client
          content:
            application/json:
              schema:
                type: object
                properties:
                  client_id: { type: string }
                  client_id_issued_at: { type: integer }
                  token_endpoint_auth_method: { type: string, enum: [none] }
                  grant_types: { type: array, items: { type: string } }
                  response_types: { type: array, items: { type: string } }
                  redirect_uris: { type: array, items: { type: string } }
                  scope: { type: string }
                  client_name: { type: string }
        "400":
          description: "invalid_redirect_uri | invalid_scope"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  error_description: { type: string }
        "502":
          description: server_error — Hydra client creation failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string, enum: [server_error] }
                  error_description: { type: string }

  # ---------------------------------------------- hydra login/consent providers
  # These are unauthenticated (pre-token) provider flows. GET routes render HTML;
  # options/verify are JSON. Documented for completeness; full request/response
  # detail lives in src/hydra/routes.ts.
...
  /api/hydra/login/verify:
    post:
      operationId: finishHydraLogin
      tags: [hydra-provider]
      security: []
      summary: Verify login assertion, accept Hydra login request
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                login_challenge: { type: string }
                challengeId: { type: string }
                credential: {}
      responses:
        "200":
          description: "{ redirectTo } on success; { error } on failure"
          content:
            application/json:
              schema:
                type: object
                properties: { redirectTo: { type: string }, error: { type: string } }
...
  /api/audit/sessions:
    get:
      operationId: listAuditSessions
      tags: [audit]
      summary: Query session-lifecycle events
      parameters:
        - { name: endpointId, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: cursor, in: query, schema: { type: string } }
        - name: limit
          in: query
          description: >-
            Page size. Clamped server-side (clampLimit): non-finite or ≤0 → 50;
            values above 500 → 500. Default 50, max 500.
          schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: Page of rows
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionLifecyclePage" }
  /api/audit/signings:
    get:
      operationId: listAuditSignings
      tags: [audit]
      summary: Query signing-request audit
      parameters:
        - {
            name: source,
            in: query,
            schema: { type: string, enum: [endpoint-auth, agent-forwarding, agent-proxy] },
          }
        - {
            name: outcome,
            in: query,
            schema: { type: string, enum: [approved, denied, expired, cancelled] },
          }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: cursor, in: query, schema: { type: string } }
        - name: limit
          in: query
          description: >-
            Page size. Clamped server-side (clampLimit): non-finite or ≤0 → 50;
            values above 500 → 500. Default 50, max 500.
          schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: Page of rows
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SigningRequestPage" }
        "400":
          description: Invalid source/outcome filter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/audit/signings/{id}:
    get:
      operationId: getAuditSigning
      tags: [audit]
      summary: Get one signing-request record
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Row
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SigningRequestRow" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ------------------------------------------------------------------ keys
...